[chassis][mgmtif][defaultroute] Fix for issue of mgmt if default route
The issue of default route on mgmt interface messing up with the default
route learned via iBGP is fixed. This fix is to increase the
administrative distance (preference) of defalut route added with
management interface as the gateway. The change is done while adding
this static route as part of staticd configuration in FRR.

Signed-off-by: vedganes <vedavinayagam.ganesan@nokia.com>