When sshd realizes that this login can't succeed due to internal device state or configuration, instead of failing right there, it proceeds to prompt for password, so as the user does not get any clue on where is the failure point.
To ensure that this login does not proceed, it replaces user provided password
with a specific pattern of characters matching length of user provided password.
This pattern is "<BS><LF><CR>>DEL>INCORRECT", which is bound to fail.

If user provided length is smaller/equal, the substring of pattern is overwritten.
If user profided length is greater, the pattern is repeated until length is exhausted.

But if the PAM-tacacs plugin would send this password to AAA, the user could get
locked out by AAA, for providing incorrect value.

Hence this fix, matches obtained password against the pattern. If match, fail before
attempting with AAA server.