[docker-mgmt-framework]: Don't overwrite /etc/passwd and /etc/group with symlinks
Because /etc/passwd and /etc/group have been overwritten with symlinks
to /host_etc/passwd and /host_etc/group, the debug container build
fails. This is because the debug container is built without /etc being
mounted at /host_etc in the container (which does happen at runtime).
Because of that, /etc/passwd and /etc/group don't exist, which causes
some package installation errors when openssh-client tries to create a
group.

This is a partial revert of 1347f29178d363b68bbb24e9e884f19e9c49bda2.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>