Fix issue of parsing syseeprom value with whitespace in middle
The platform api for parsing syseeprom information read from STATE DB has issue
with parsing the value part that has whitespace in the middle. The current
code assumes that the value part does not have whitespace. So everything after
the whitespace will be ignored. The syseeprom values returned from platform
API do not match the output of "show platform syseeprom".

This change improved the regular expression for parsing syseeprom values to
accommodate whitespaces in the value.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>