Align entry name in copp_cfg.j2 with feature name
There is a logic in copp manager which loads copp items only if the feature is enabled.
It requires the feature name aligns with copp item name otherwise the copp item can not be loaded.
Currently, dhcp copp item is used by feature dhcp_relay but the name doesn't align.
This is to fix it.

Signed-off-by: Stephen Sun <stephens@nvidia.com>