Sonic image build optimization
The current SONIC build infrastructure always prepares the rootfs for final image generation.
Because of rootfs preparation, final image generation takes long regardless of SONiC changes.

The rootfs preparation consists of two parts.
	Debian packages installation
		Bootstrap preparation
		General packages installation, such as curl, vim, sudo, python3, etc
	Sonic packages installation
		Packages that are built and installed from the sonic repo.
		Docker images that are built and installed from the sonic repo

The build time can be optimized by generating the Debian packages as a base image and
it can be run in parallel with the other targets, before the final image.
	Benefits:
		- High hit rate, for fewer dependencies.
		- Reduce the cache size.
		- Improve the concurrency when the cache is not hit, the step has small dependencies and can be run with any other steps.

Other enhancements:
	- The docker load is also optimized by running in parallel.
	- Added support for gipz compression
	- Added tmpfs support for rootfs and final image generation.