Fix Celestica Seastone2 LED init script to not break port status
Fixes #10103.

The LED init script that was added in #9769 enables the port LEDs, but
it turns out that the linkscan change combined with Broadcom SAI 6.0
breaks detection of port status on boot.

Without the linkscan and sleep, the LED still work as expected, and
the port status is correct again.

Signed-off-by: Lukas Stockner <lstockner@genesiscloud.com>