[write_standby] update write_standby.py script
The initial value has to be present for the state machines to work.
In active-standby dual-tor scenario, or any hardware mux scenario,
the value will be updtaed eventually with a delay.

However, in active-active dual-tor scenario, there is no other
mechanism to initialize the value and get state machines started.
So this script will have to write something at start up time.

For active-active dualtor, 'active' is a more preferred initial
value, the state machine will switch the state to standby soon if
link prober found link not in good state.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>