03.Version-cache - framework environment settings
It defines and passes the necessary version cache environment variables
to the caching framework.

It adds the utils script for shared cache file access.

It also adds the post-cleanup logic for cleaning the unwanted files from
the docker/image after the version cache creation.

The following variable controls the version caching feature. By default, it is turned off.
    SONIC_VERSION_CACHE_METHOD = [none/cache]
    SONIC_VERSION_CACHE_SOURCE = /path/to/local/storage