Fix dhcp-relay and macsec dockers not being marked as local packages
When INSTALL_DEBUG_TOOLS=y is set (i.e. debug tools and containers are
enabled), the non-debug versions of the dhcp-relay and macsec dockers
are not added into SONIC_PACKAGES_LOCAL, and thus the tagging for those
packages is not handled correctly when loading those dockers.

To fix this, change the code that checks to see if something is part of
a local package by searching for both just the package name and with the
-dbg suffix to the name.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>