sonic-host-services: Pin deepdiff to version 6.2.2
The deepdiff python package was recently updated to 6.2.3. As part of
this, a dependency was introduced on orjson. There's no armv8l python
wheel available for orjson, which means it needs to be built from
source. However, building it requires rust (which Buster and Bullseye
don't have a new enough version of) and maturin.

As a quick fix, pin this to version 6.2.2, before the orjson dependency
is introduced.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>