Don't create the members@ array in config_db for PC when reading from minigraph
When loading from minigraph, for port channels, don't create the
members@ array in config_db in the PORTCHANNEL table. This is no longer
needed or used.

In addition, when adding a port channel member from the CLI, that member
doesn't get added into the members@ array, resulting in a bit of
inconsistency. This gets rid of that inconsistency.

Fixes #11873.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>