[Arista] Disable ATA NCQ for a few products
Some products might experience an occasional IO failure in the
communication between CPU and SSD.

Based on some research it could be attributable to some device not
handling ATA NCQ (Native Command Queue).

This change disable NCQ on the affected drive for a set of products.

When the fix is applied, these 2 patterns can be found in the dmesg.
ata1.00: FORCE: horkage modified (noncq)
NCQ (not used)