frrcfgd: fix incorrect rendering of route-map references
A reference to a route-maps under a BGP address family is a single-value
string, not a multi-value list. Looping over it yields incorrect
configuration being rendered, for example:

```
  neighbor peergroup route-map R in
  neighbor peergroup route-map M in
  neighbor peergroup route-map A in
  neighbor peergroup route-map P in
```

These overwrite each other as the configuration is loaded, making the
effective FRR contain only the last one listed. This route-map (in this
case called `P`) does most likely not exist, leading to an effective
configuration which rejects all routes.

The correct and expected rendered configuration would be:

```
  neighbor peergroup route-map RMAP in
```

Signed-off-by: Tore Anderson <tore@redpill-linpro.com>